3214 - getRepositoryItems => TablePaginatedActions.getData
