--- src/client/store/ui/repository/reducers/repositoryItemsReducer.ts
@@ -1,10 +1,0 @@
-import { ActionReducerMapBuilder } from '@reduxjs/toolkit'
-
-import { getRepositoryItems } from 'client/store/ui/repository/actions/getRepositoryItems'
-import { RepositoryState } from 'client/store/ui/repository/state'
-
-export const repositoryItemsReducer = (builder: ActionReducerMapBuilder<RepositoryState>) => {
-  builder.addCase(getRepositoryItems.fulfilled, (state, { payload }) => {
-    state.repositoryItems = payload
-  })
-}
