+++ src/client/store/ui/repository/selectors/index.ts
@@ -6,12 +6,10 @@ import { RepositorySlice } from 'client/store/ui/repository/slice'
 const _getState = (state: RootState) => state.ui[RepositorySlice.name]
 const isLoading = createSelector(_getState, (repository) => repository.loading)
 const getRepositoryItem = createSelector(_getState, (repository) => repository.repositoryItem)
-const getRepositoryItems = createSelector(_getState, (repository) => repository.repositoryItems)
 const getRepositoryItemValidation = createSelector(_getState, (repository) => repository.repositoryItemValidation)
 
 export const RepositorySelectors = {
   isLoading,
   getRepositoryItem,
-  getRepositoryItems,
   getRepositoryItemValidation,
 }
