+++ src/client/store/ui/repository/state.ts
@@ -4,7 +4,6 @@ import { RepositoryItemValidation } from 'meta/cycleData/repository'
 export type RepositoryState = {
   loading: boolean
   repositoryItem?: Partial<RepositoryItem>
-  repositoryItems?: Array<RepositoryItem>
   repositoryItemValidation?: RepositoryItemValidation
 }
